requirements.txt
django>=5.0.3,<5.1.0
djangorestframework
django-cors-headers
google-generativeai
python-dotenv 
djangorestframework-simplejwt